Bug 966963: Remove unnecessary versioned conf files from php cart
The php 5.3/5.4 configuration files all symlink back to shared. Remove the specific version
directories/files and use only the shared directory for setup. This also fixes a bug where
the symlinks to shared become broken during setup (as ERBs are removed during processing)
and cause the remaining symlinks to be re-processed as ERBs (which fails).